Removing what will end up being duplicated code
 - Using generics to generalize the import flow with an importSet type.
