+++ data/boltstorage.go
@@ -205,146 +205,6 @@ func (b *boltStorage) Export(dir string) (outName string, err error) {
 	return outName, nil
 }
 
-func (b *boltStorage) Import(file string, merge MergeStrategy) error {
-	in, err := os.Open(file)
-	if err != nil {
-		return err
-	}
-	defer func() {
-		_ = in.Close()
-	}()
-
-	var model exportModel
-	if err := json.NewDecoder(in).Decode(&model); err != nil {
-		return err
-	}
-
-	tx, err := b.store.Bolt().Begin(true)
-	if err != nil {
-		return err
-	}
-	rollback := func(err error) error {
-		if rerr := tx.Rollback(); rerr != nil {
-			return fmt.Errorf("%w: %v", err, rerr)
-		}
-		return err
-	}
-
-	if err := b.importTasks(tx, model, merge); err != nil {
-		return rollback(err)
-	}
-	if err := b.importTimeEntries(tx, model, merge); err != nil {
-		return rollback(err)
-	}
-
-	if err := tx.Commit(); err != nil {
-		return rollback(err)
-	}
-	return nil
-}
-
-func (b *boltStorage) importTasks(tx *bbolt.Tx, model exportModel, merge MergeStrategy) error {
-
-	for _, imported := range model.Tasks {
-		switch {
-		case imported.ID == 0:
-			return fmt.Errorf("%w: %v", ErrUnmappedReqdImportField, "missing/unset ID field")
-		case len(imported.Name) == 0:
-			return fmt.Errorf("%w: %v", ErrUnmappedReqdImportField, "missing/empty name field")
-		}
-
-		var (
-			dupe Task
-			err  error
-		)
-		err = b.store.TxFindOne(tx, &dupe, bolthold.Where(bolthold.Key).Eq(imported.ID))
-		if err != nil {
-			if err != bolthold.ErrNotFound {
-				return err
-			}
-
-			// No dupe found, import it directly.
-			if err := b.store.TxInsert(tx, imported.ID, imported); err != nil {
-				return err
-			}
-			continue
-		}
-
-		// Check if the dupe represents a different state.
-		if dupe == imported {
-			continue
-		}
-
-		// A duplicate was found, employ the MergeStrategy.
-		switch merge {
-		case MergeOverwrite:
-			err := b.store.TxUpdate(tx, dupe.ID, imported)
-			if err != nil {
-				return err
-			}
-			continue
-		case MergeKeepInternal:
-			continue
-		case MergeError:
-			return errors.New("conflict discovered, rolling back import")
-		default:
-			return fmt.Errorf("unrecognized merge strategy '%s'", merge)
-		}
-	}
-	return nil
-}
-
-func (b *boltStorage) importTimeEntries(tx *bbolt.Tx, model exportModel, merge MergeStrategy) error {
-
-	for _, imported := range model.TimeEntries {
-		switch {
-		case imported.ID == 0:
-			return fmt.Errorf("%w: %v", ErrUnmappedReqdImportField, "missing/unset ID field")
-		case imported.Start == time.Time{}:
-			return fmt.Errorf("%w: %v", ErrUnmappedReqdImportField, "missing/empty start field")
-		}
-
-		var (
-			dupe TimeEntry
-			err  error
-		)
-		err = b.store.TxFindOne(tx, &dupe, bolthold.Where(bolthold.Key).Eq(imported.ID))
-		if err != nil {
-			if err != bolthold.ErrNotFound {
-				return err
-			}
-
-			// No dupe found, import it directly.
-			if err := b.store.TxInsert(tx, imported.ID, imported); err != nil {
-				return err
-			}
-			continue
-		}
-
-		// Check if the dupe represents a different state.
-		if dupe == imported {
-			continue
-		}
-
-		// A duplicate was found, employ the MergeStrategy.
-		switch merge {
-		case MergeOverwrite:
-			err := b.store.TxUpdate(tx, dupe.ID, imported)
-			if err != nil {
-				return err
-			}
-			continue
-		case MergeKeepInternal:
-			continue
-		case MergeError:
-			return errors.New("conflict discovered, rolling back import")
-		default:
-			return fmt.Errorf("unrecognized merge strategy '%s'", merge)
-		}
-	}
-	return nil
-}
-
 func (b *boltStorage) StartTimeEntry(taskID uint64) (TimeEntry, error) {
 	var entry TimeEntry
 	err := b.store.Bolt().Update(func(tx *bbolt.Tx) error {
